requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-docx==1.1.0

